Refactor config, logging packages
- create `internal/config` package from app-specific config
  init code
- merge `internal/logging` package to `internal/config`
- dedupe logger structured fields

refs GH-1
