+++ cmd/check_cert/main.go
@@ -9,31 +9,20 @@ package main
 
 import (
 	"crypto/tls"
+	"errors"
 	"fmt"
 	"net"
-	"os"
 	"strings"
 	"time"
 
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
+	zlog "github.com/rs/zerolog/log"
 
 	"github.com/atc0005/check-certs/internal/certs"
-	"github.com/atc0005/check-certs/internal/logging"
+	"github.com/atc0005/check-certs/internal/config"
 	"github.com/atc0005/go-nagios"
 )
 
 func main() {
-
-	// Setup configuration by parsing user-provided flags
-	config := Config{}
-	config.handleFlagsConfig()
-
-	// Display application info and exit
-	if config.ShowVersion {
-		fmt.Println(Version())
-		os.Exit(0)
-	}
 
 	// Set initial "state" as valid, adjust as we go.
 	var nagiosExitState = nagios.ExitState{
@@ -44,81 +33,59 @@ func main() {
 	// defer this from the start so it is the last deferred function to run
 	defer nagiosExitState.ReturnCheckResults()
 
-	if err := config.Validate(); err != nil {
+	// Setup configuration by parsing user-provided flags.
+	isPlugin := true
+	cfg, cfgErr := config.New(isPlugin)
+	switch {
+	case errors.Is(cfgErr, config.ErrVersionRequested):
+		fmt.Println(config.Version())
+
+		return
+
+	case cfgErr != nil:
+		// We're using the standalone Err function from rs/zerolog/log as we
+		// do not have a working configuration.
+		zlog.Err(cfgErr).Msg("Error initializing application")
 		nagiosExitState.ServiceOutput = fmt.Sprintf(
-			"%s: Error validating configuration",
+			"%s: Error initializing application",
 			nagios.StateCRITICALLabel,
 		)
-		nagiosExitState.LastError = err
+		nagiosExitState.LastError = cfgErr
 		nagiosExitState.ExitStatusCode = nagios.StateCRITICALExitCode
-		log.Err(err).Msg("Error validating configuration")
 
-		// no need to go any further, we *want* to exit right away; we don't
-		// have a working configuration and there isn't anything further to do
 		return
 	}
 
-	if config.EmitBranding {
+	if cfg.EmitBranding {
 		// If enabled, show application details at end of notification
-		nagiosExitState.BrandingCallback = Branding("Notification generated by ")
+		nagiosExitState.BrandingCallback = config.Branding("Notification generated by ")
 	}
 
 	// Use provided threshold values to calculate the expiration times that
 	// should trigger either a WARNING or CRITICAL state.
 	now := time.Now().UTC()
-	certsExpireAgeWarning := now.AddDate(0, 0, config.AgeWarning)
-	certsExpireAgeCritical := now.AddDate(0, 0, config.AgeCritical)
+	certsExpireAgeWarning := now.AddDate(0, 0, cfg.AgeWarning)
+	certsExpireAgeCritical := now.AddDate(0, 0, cfg.AgeCritical)
 
 	nagiosExitState.WarningThreshold = fmt.Sprintf(
 		"Expires before %v (%d days)",
 		certsExpireAgeWarning.Format(certs.CertValidityDateLayout),
-		config.AgeWarning,
+		cfg.AgeWarning,
 	)
 	nagiosExitState.CriticalThreshold = fmt.Sprintf(
 		"Expires before %v (%d days)",
 		certsExpireAgeCritical.Format(certs.CertValidityDateLayout),
-		config.AgeCritical,
+		cfg.AgeCritical,
 	)
 
-	// Note: Nagios doesn't look at stderr, only stdout. We have to make sure
-	// that only whatever output is meant for consumption is emitted to stdout
-	// and whatever is meant for troubleshooting is sent to stderr. To help
-	// keep these two goals separate (and because Nagios doesn't really do
-	// anything special with JSON output from plugins), we use stdlib fmt
-	// package output functions for Nagios via stdout and logging package for
-	// troubleshooting via stderr.
-	//
-	// Also, set common fields here so that we don't have to repeat them
-	// explicitly later. This will hopefully help to standardize the log
-	// messages to make them easier to search through later when
-	// troubleshooting.
-	log := zerolog.New(os.Stderr).With().Caller().
-		Str("version", version).
-		Str("logging_level", config.LoggingLevel).
-		Str("server", config.Server).
-		Int("port", config.Port).
-		Int("age_warning", config.AgeWarning).
-		Int("age_critical", config.AgeCritical).
-		Str("expected_sans_entries", config.SANsEntries.String()).Logger()
+	log := cfg.Log.With().
+		Str("expected_sans_entries", cfg.SANsEntries.String()).
+		Logger()
 
-	if err := logging.SetLoggingLevel(config.LoggingLevel); err != nil {
-		nagiosExitState.LastError = err
-		nagiosExitState.ServiceOutput = fmt.Sprintf(
-			"%s: Error configuring logging level",
-			nagios.StateCRITICALLabel,
-		)
-		nagiosExitState.ExitStatusCode = nagios.StateCRITICALExitCode
-		log.Err(err).Msg("configuring logging level")
-
-		// no need to defer this, we *want* to exit right away; we don't have
-		// a working configuration and there isn't anything further to do
-		return
-	}
-
-	server := fmt.Sprintf("%s:%d", config.Server, config.Port)
+	server := fmt.Sprintf("%s:%d", cfg.Server, cfg.Port)
 
 	log.Debug().Msg("Connecting to remote server")
-	cfg := tls.Config{
+	tlsConfig := tls.Config{
 		// Allow insecure connection so that we can check not only the initial
 		// certificate (which may be expired), but others in the chain also to
 		// potentially catch any intermediates which may also be expired.
@@ -129,10 +96,10 @@ func main() {
 
 	// Create custom dialer with user-specified timeout value
 	dialer := &net.Dialer{
-		Timeout: time.Duration(config.Timeout) * time.Second,
+		Timeout: time.Duration(cfg.Timeout) * time.Second,
 	}
 
-	conn, connErr := tls.DialWithDialer(dialer, "tcp", server, &cfg)
+	conn, connErr := tls.DialWithDialer(dialer, "tcp", server, &tlsConfig)
 	if connErr != nil {
 		nagiosExitState.LastError = connErr
 		nagiosExitState.ServiceOutput = fmt.Sprintf(
@@ -141,7 +108,7 @@ func main() {
 			server,
 		)
 		nagiosExitState.ExitStatusCode = nagios.StateCRITICALExitCode
-		log.Error().Err(connErr).Str("server", server).Msg("error connecting to server")
+		log.Error().Err(connErr).Msg("error connecting to server")
 
 		// no need to go any further, we *want* to exit right away; we don't
 		// have a connection to the remote server and there isn't anything
@@ -175,19 +142,19 @@ func main() {
 			server,
 		)
 		nagiosExitState.ExitStatusCode = nagios.StateCRITICALExitCode
-		log.Error().Err(noCertsErr).Str("server", server).Msg("no certificates found")
+		log.Error().Err(noCertsErr).Msg("no certificates found")
 
 		return
 	}
 
 	if certsSummary.TotalCertsCount > 0 {
 
-		hostnameValue := config.Server
+		hostnameValue := cfg.Server
 
 		// Allow user to explicitly specify which hostname should be used
 		// for comparison against the leaf certificate.
-		if config.DNSName != "" {
-			hostnameValue = config.DNSName
+		if cfg.DNSName != "" {
+			hostnameValue = cfg.DNSName
 		}
 
 		// Verify leaf certificate is valid for the provided server FQDN; we
@@ -220,8 +187,7 @@ func main() {
 			nagiosExitState.ExitStatusCode = nagios.StateCRITICALExitCode
 			log.Error().
 				Err(err).
-				Str("server", config.Server).
-				Str("dns_name", config.DNSName).
+				Str("hostname", hostnameValue).
 				Str("cert_cn", certChain[0].Subject.CommonName).
 				Str("sans_entries", fmt.Sprintf("%s", certChain[0].DNSNames)).
 				Msg("hostname does not match first cert in chain")
@@ -231,20 +197,20 @@ func main() {
 		}
 
 		log.Debug().
-			Str("hostname", config.Server).
+			Str("hostname", hostnameValue).
 			Str("cert_cn", certChain[0].Subject.CommonName).
 			Msg("provided hostname %q matches server certificate")
 
 	}
 
 	// check SANS entries if provided via command-line
-	if len(config.SANsEntries) > 0 {
+	if len(cfg.SANsEntries) > 0 {
 
 		// Check for special keyword, skip SANs entry checks if provided
-		firstSANsEntry := strings.ToLower(strings.TrimSpace(config.SANsEntries[0]))
-		if firstSANsEntry != strings.ToLower(strings.TrimSpace(SkipSANSCheckKeyword)) {
+		firstSANsEntry := strings.ToLower(strings.TrimSpace(cfg.SANsEntries[0]))
+		if firstSANsEntry != strings.ToLower(strings.TrimSpace(config.SkipSANSCheckKeyword)) {
 
-			if mismatched, err := certs.CheckSANsEntries(certChain[0], config.SANsEntries); err != nil {
+			if mismatched, err := certs.CheckSANsEntries(certChain[0], cfg.SANsEntries); err != nil {
 
 				nagiosExitState.LastError = err
 
@@ -258,13 +224,13 @@ func main() {
 					"%s: Mismatch of %d SANs entries for certificate %q",
 					nagios.StateCRITICALLabel,
 					mismatched,
-					config.Server,
+					cfg.Server,
 				)
 
 				nagiosExitState.ExitStatusCode = nagios.StateWARNINGExitCode
 				log.Warn().
 					Err(nagiosExitState.LastError).
-					Int("sans_entries_requested", len(config.SANsEntries)).
+					Int("sans_entries_requested", len(cfg.SANsEntries)).
 					Int("sans_entries_found", len(certChain)).
 					Msg("SANs entries mismatch")
