+++ internal/config/config.go
@@ -1,0 +1,175 @@
+// Copyright 2020 Adam Chalkley
+//
+// https://github.com/atc0005/check-cert
+//
+// Licensed under the MIT License. See LICENSE file in the project root for
+// full license information.
+
+package config
+
+import (
+	"errors"
+	"flag"
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/rs/zerolog"
+)
+
+// Updated via Makefile builds. Setting placeholder value here so that
+// something resembling a version string will be provided for non-Makefile
+// builds.
+var version string = "x.y.z"
+
+// ErrVersionRequested indicates that the user requested application version
+// information.
+var ErrVersionRequested = errors.New("version information requested")
+
+// multiValueFlag is a custom type that satisfies the flag.Value interface in
+// order to accept multiple values for some of our flags.
+type multiValueFlag []string
+
+// String returns a comma separated string consisting of all slice elements.
+func (i *multiValueFlag) String() string {
+
+	// From the `flag` package docs:
+	// "The flag package may call the String method with a zero-valued
+	// receiver, such as a nil pointer."
+	if i == nil {
+		return ""
+	}
+
+	return strings.Join(*i, ", ")
+}
+
+// Set is called once by the flag package, in command line order, for each
+// flag present.
+func (i *multiValueFlag) Set(value string) error {
+
+	// split comma-separated string into multiple values, toss whitespace
+	items := strings.Split(value, ",")
+	for index, item := range items {
+		items[index] = strings.TrimSpace(item)
+	}
+
+	// add them to the collection
+	*i = append(*i, items...)
+	return nil
+}
+
+// Config represents the application configuration as specified via
+// command-line flags.
+type Config struct {
+
+	// SANsEntries is the list of Subject Alternate Names (SANs) to verify are
+	// present on the examined certificate. This value is provided a
+	// comma-separated list.
+	SANsEntries multiValueFlag
+
+	// Filename is the fully-qualified path to a file containing one or more
+	// certificates.
+	Filename string
+
+	// Server is the fully-qualified domain name of the system running a
+	// certificate-enabled service.
+	Server string
+
+	// DNSName is the fully-qualified domain name associated with the
+	// certificate. This is usually specified when the FQDN or IP used to make
+	// the connection is different than the Common Name or Subject Alternate
+	// Names entries associated with the certificate.
+	DNSName string
+
+	// Port is the TCP port used by the certifcate-enabled service.
+	Port int
+
+	// LoggingLevel is the supported logging level for this application.
+	LoggingLevel string
+
+	// AgeWarning is the number of days remaining before certificate
+	// expiration when this application will flag the NotAfter certificate
+	// field as a WARNING state.
+	AgeWarning int
+
+	// AgeCritical is the number of days remaining before certificate
+	// expiration when this application will flag the NotAfter certificate
+	// field as a CRITICAL state.
+	AgeCritical int
+
+	// Timeout is the number of seconds allowed before the connection attempt
+	// to a remote certificate-enabled service is abandoned and an error
+	// returned.
+	Timeout int
+
+	// EmitBranding controls whether "generated by" text is included at the
+	// bottom of application output. This output is included in the Nagios
+	// dashboard and notifications. This output may not mix well with branding
+	// output from other tools such as atc0005/send2teams which also insert
+	// their own branding output.
+	EmitBranding bool
+
+	// EmitCertText controls whether x509 TLS certificates are printed to
+	// stdout using an OpenSSL-inspired text format. There is a good bit of
+	// output text, so this setting defaults to false.
+	EmitCertText bool
+
+	// ShowVersion is a flag indicating whether the user opted to display only
+	// the version string and then immediately exit the application.
+	ShowVersion bool
+
+	// Log is an embedded zerolog Logger initialized via config.New().
+	Log zerolog.Logger
+}
+
+// Usage is a custom override for the default Help text provided by the flag
+// package. Here we prepend some additional metadata to the existing output.
+var Usage = func() {
+	fmt.Fprintln(flag.CommandLine.Output(), "\n"+Version()+"\n")
+	fmt.Fprintf(flag.CommandLine.Output(), "Usage of %s:\n", os.Args[0])
+	flag.PrintDefaults()
+}
+
+// Version emits application name, version and repo location.
+func Version() string {
+	return fmt.Sprintf("%s %s (%s)", myAppName, version, myAppURL)
+}
+
+// Branding accepts a message and returns a function that concatenates that
+// message with version information. This function is intended to be called as
+// a final step before application exit after any other output has already
+// been emitted.
+func Branding(msg string) func() string {
+	return func() string {
+		return strings.Join([]string{msg, Version()}, "")
+	}
+}
+
+// New is a factory function that produces a new Config object based on user
+// provided flag and config file values. It is responsible for validating
+// user-provided values and initializing the logging settings used by this
+// application.
+func New(isPlugin bool) (*Config, error) {
+	var config Config
+
+	config.handleFlagsConfig(isPlugin)
+
+	if config.ShowVersion {
+		return nil, ErrVersionRequested
+	}
+
+	if err := config.validate(isPlugin); err != nil {
+		return nil, fmt.Errorf("configuration validation failed: %w", err)
+	}
+
+	// initialize logging just as soon as validation is complete
+	if err := config.setupLogging(isPlugin); err != nil {
+		return nil, fmt.Errorf(
+			"failed to set logging configuration: %w",
+			err,
+		)
+	}
+
+	return &config, nil
+
+}
