+++ internal/config/doc.go
@@ -1,0 +1,10 @@
+// Copyright 2020 Adam Chalkley
+//
+// https://github.com/atc0005/check-cert
+//
+// Licensed under the MIT License. See LICENSE file in the project root for
+// full license information.
+
+// Package config provides types and functions to collect, validate and apply
+// user-provided settings.
+package config
