--- internal/logging/doc.go
@@ -1,11 +1,0 @@
-// Copyright 2020 Adam Chalkley
-//
-// https://github.com/atc0005/check-cert
-//
-// Licensed under the MIT License. See LICENSE file in the project root for
-// full license information.
-
-// Package logging provides shared constants and a map generated at
-// initialization between flag-provided logging levels and internal logging
-// levels used by the logging package applications in this module rely on.
-package logging
